JSConfig to easly import without ../../.. . Fixed performance removing call for guild-stats endpoint three times, now calling just once
