+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "tibia-bosses-tracker",
-  "version": "0.1.0",
-  "private": true,
+  "version": "1.14.0",
   "dependencies": {
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
@@ -12,15 +11,13 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.4.0",
-    "cheerio": "^1.0.0-rc.12",
     "date-fns": "^3.3.1",
     "firebase": "^9.23.0",
-    "json-loader": "^0.5.7",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router": "^6.14.0",
     "react-router-dom": "^6.13.0",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "react-spinners": "^0.13.8",
     "use-debounce": "^9.0.4",
     "web-vitals": "^2.1.4"
