+++ src/components/main-app-bar/index.js
@@ -8,7 +8,7 @@ import {
 } from "@material-ui/core";
 import CloseIcon from "@material-ui/icons/Close";
 import MenuOpenIcon from "@material-ui/icons/MenuOpen";
-import { useFilterContext } from "../../context/FilterContext";
+import { useFilterContext } from "context/FilterContext";
 import { FilterList } from "@material-ui/icons";
 import { Link } from "react-router-dom";
 import NewReleasesIcon from "@material-ui/icons/NewReleases";
